requirements.txt
streamlit
pdfplumber
python-docx
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-groq
chromadb
sentence-transformers
huggingface-hub
cryptography
streamlit-cookies-manager
python-dotenv